Fixing return types for finance information

+++ Apps.XTRF/Responses/FinanceInformation.cs
@@ -9,12 +9,12 @@ namespace Apps.XTRF.Responses
     public class FinanceInformation
     {
         public int CurrencyId { get; set; }
-        public int TotalAgreed { get; set; }
+        public double TotalAgreed { get; set; }
         public double TotalCost { get; set; }
         public double Profit { get; set; }
         public double Margin { get; set; }
         public IEnumerable<Payable> Payables { get; set; }
         public IEnumerable<Receivable> Receivables { get; set; }
-        public int ROI { get; set; }
+        public double ROI { get; set; }
     }
 }
